grouped up far samples
