+++ samples/CMakeLists.txt
@@ -67,6 +67,7 @@ set(S2D_SAMPLES
 	collection/human.h
 	collection/sample_contact.cpp
 	collection/sample_joints.cpp
+	collection/sample_far.cpp
 )
 
 add_executable(samples ${S2D_SAMPLES})
