+++ samples/collection/sample_contact.cpp
@@ -797,39 +797,30 @@ public:
 		m_row = 0;
 		m_column = 0;
 		m_count = 0;
-	}
-
-	void CreateCircle()
-	{
-		float x = -9.0f + m_column * 18.0f / e_gridCount;
-		float y = 1.0f + m_row * 18.0f / e_gridCount;
 
 		s2BodyDef bodyDef = s2_defaultBodyDef;
 		bodyDef.type = s2_dynamicBody;
-		bodyDef.position = {x, y};
 		bodyDef.gravityScale = 0.0f;
 
-		s2BodyId bodyId = s2CreateBody(m_worldId, &bodyDef);
-
 		s2Circle circle = {{0.0f, 0.0f}, 0.5f};
-		s2CreateCircleShape(bodyId, &s2_defaultShapeDef, &circle);
-	}
 
-	void Step(Settings& settings, s2Color bodyColor) override
-	{
-		if (settings.timeStep > 0.0f && m_count < e_maxCount && m_stepCount % 1 == 0)
+		while (m_count < e_maxCount)
 		{
-			CreateCircle();
-			m_count += 1;
-			m_row += 1;
-			if (m_row == e_gridCount)
+			m_row = 0;
+			for (int i = 0; i < e_gridCount; ++i)
 			{
-				m_row = 0;
-				m_column += 1;
-			}
-		}
+				float x = -8.75f + m_column * 18.0f / e_gridCount;
+				float y = 1.5f + m_row * 18.0f / e_gridCount;
 
-		Sample::Step(settings, bodyColor);
+				bodyDef.position = {x, y};
+				s2BodyId bodyId = s2CreateBody(m_worldId, &bodyDef);
+				s2CreateCircleShape(bodyId, &s2_defaultShapeDef, &circle);
+
+				m_count += 1;
+				m_row += 1;
+			}
+			m_column += 1;
+		}
 	}
 
 	static Sample* Create(const Settings& settings, s2SolverType solverType)
@@ -843,205 +834,3 @@ public:
 };
 
 static int sampleConfined = RegisterSample("Contact", "Confined", Confined::Create);
-
-// A pyramid far from the origin
-class FarPyramid : public Sample
-{
-public:
-	FarPyramid(const Settings& settings, s2SolverType solverType)
-		: Sample(settings, solverType)
-	{
-		// s2Vec2 origin = {0.0f, 0.0f};
-		s2Vec2 origin = {50000.0f, -45000.0f};
-		//s2Vec2 origin = {900000.0f, -800000.0f};
-
-		float originx = 65000.0f;
-		float nextfloat = nextafterf(originx, 2.0f * originx);
-		float ulp = nextfloat - originx;
-		printf("ulp = %g\n", ulp);
-
-		if (settings.restart == false)
-		{
-			g_camera.m_center = s2Add({0.0f, 6.0f}, origin);
-			g_camera.m_zoom = 0.3f;
-		}
-
-		{
-			s2BodyDef bodyDef = s2_defaultBodyDef;
-			bodyDef.position = s2Add({0.0f, -1.0f}, origin);
-			s2BodyId groundId = s2CreateBody(m_worldId, &bodyDef);
-
-			s2Polygon box = s2MakeBox(100.0f, 1.0f);
-			s2ShapeDef shapeDef = s2_defaultShapeDef;
-			s2CreatePolygonShape(groundId, &shapeDef, &box);
-		}
-
-		s2BodyDef bodyDef = s2_defaultBodyDef;
-		bodyDef.type = s2_dynamicBody;
-
-		s2ShapeDef shapeDef = s2_defaultShapeDef;
-		shapeDef.density = 1.0f;
-
-		int baseCount = 10;
-
-		float h = 0.5f;
-		s2Polygon box = s2MakeSquare(h);
-
-		float shift = 1.25f * h;
-
-		for (int i = 0; i < baseCount; ++i)
-		{
-			float y = (2.0f * i + 1.0f) * shift + 0.5f;
-
-			for (int j = i; j < baseCount; ++j)
-			{
-				float x = (i + 1.0f) * shift + 2.0f * (j - i) * shift - h * baseCount;
-
-				bodyDef.position = s2Add({x, y}, origin);
-
-				s2BodyId bodyId = s2CreateBody(m_worldId, &bodyDef);
-				s2CreatePolygonShape(bodyId, &shapeDef, &box);
-			}
-		}
-	}
-
-	static Sample* Create(const Settings& settings, s2SolverType solverType)
-	{
-		return new FarPyramid(settings, solverType);
-	}
-};
-
-static int sampleFarPyramid = RegisterSample("Contact", "Far Pyramid", FarPyramid::Create);
-
-class FarStack : public Sample
-{
-public:
-	FarStack(const Settings& settings, s2SolverType solverType)
-		: Sample(settings, solverType)
-	{
-		//s2Vec2 origin = {0.0f, 0.0f};
-		s2Vec2 origin = {40000.0f, -25000.0f};
-		//s2Vec2 origin = {900000.0f, -800000.0f};
-
-		if (settings.restart == false)
-		{
-			g_camera.m_center = s2Add({0.0f, 1.0f}, origin);
-			g_camera.m_zoom = 0.08f;
-		}
-
-		{
-			s2BodyDef bodyDef = s2_defaultBodyDef;
-			bodyDef.position = s2Add({0.0f, -1.0f}, origin);
-			s2BodyId groundId = s2CreateBody(m_worldId, &bodyDef);
-
-			s2Polygon box = s2MakeBox(10.0f, 1.0f);
-			s2ShapeDef shapeDef = s2_defaultShapeDef;
-			s2CreatePolygonShape(groundId, &shapeDef, &box);
-		}
-
-		s2BodyDef bodyDef = s2_defaultBodyDef;
-		bodyDef.type = s2_dynamicBody;
-
-		s2ShapeDef shapeDef = s2_defaultShapeDef;
-		shapeDef.density = 1.0f;
-
-		{
-			bodyDef.position = s2Add({1.875f, 0.125f}, origin);
-			s2BodyId bodyId = s2CreateBody(m_worldId, &bodyDef);
-			s2Circle circle = {{0.0f, 0.0f}, 0.1f};
-			s2CreateCircleShape(bodyId, &shapeDef, &circle);
-		}
-
-		{
-			bodyDef.position = s2Add({-1.875f, 0.15f}, origin);
-			s2BodyId bodyId = s2CreateBody(m_worldId, &bodyDef);
-			s2Polygon box = s2MakeBox(0.1f, 0.125f);
-			s2CreatePolygonShape(bodyId, &shapeDef, &box);
-		}
-
-		{
-			bodyDef.position = s2Add({0.0f, 0.325f}, origin);
-			s2BodyId bodyId = s2CreateBody(m_worldId, &bodyDef);
-			s2Polygon box = s2MakeBox(2.0f, 0.05f);
-			s2CreatePolygonShape(bodyId, &shapeDef, &box);
-		}
-
-		{
-			bodyDef.position = s2Add({-0.5f, 0.9f}, origin);
-			s2BodyId bodyId = s2CreateBody(m_worldId, &bodyDef);
-			s2Polygon box = s2MakeSquare(0.25f);
-			s2CreatePolygonShape(bodyId, &shapeDef, &box);
-		}
-
-		{
-			bodyDef.position = s2Add({-0.55f, 1.7f}, origin);
-			s2BodyId bodyId = s2CreateBody(m_worldId, &bodyDef);
-			s2Polygon box = s2MakeSquare(0.5f);
-			s2CreatePolygonShape(bodyId, &shapeDef, &box);
-		}
-	}
-
-	static Sample* Create(const Settings& settings, s2SolverType solverType)
-	{
-		return new FarStack(settings, solverType);
-	}
-};
-
-static int sampleFarStack = RegisterSample("Contact", "Far Stack", FarStack::Create);
-
-class FarRecovery : public Sample
-{
-public:
-	FarRecovery(const Settings& settings, s2SolverType solverType)
-		: Sample(settings, solverType)
-	{
-		// s2Vec2 origin = {0.0f, 0.0f};
-		s2Vec2 origin = {40000.0f, -35000.0f};
-		//s2Vec2 origin = {900000.0f, -800000.0f};
-
-		if (settings.restart == false)
-		{
-			g_camera.m_center = s2Add({0.0f, 5.0f}, origin);
-			g_camera.m_zoom = 0.25f;
-		}
-
-		int baseCount = 4;
-		float overlap = 0.25f;
-		float extent = 0.5f;
-
-		s2BodyDef bodyDef = s2_defaultBodyDef;
-		bodyDef.position = origin;
-		s2BodyId groundId = s2CreateBody(m_worldId, &bodyDef);
-		s2Segment segment = {{-40.0f, 0.0f}, {40.0f, 0.0f}};
-		s2CreateSegmentShape(groundId, &s2_defaultShapeDef, &segment);
-
-		bodyDef.type = s2_dynamicBody;
-
-		s2Polygon box = s2MakeSquare(extent);
-
-		float fraction = 1.0f - overlap;
-		float y = extent;
-		for (int i = 0; i < baseCount; ++i)
-		{
-			float x = fraction * extent * (i - baseCount);
-			for (int j = i; j < baseCount; ++j)
-			{
-				bodyDef.position = s2Add({x, y}, origin);
-				s2BodyId bodyId = s2CreateBody(m_worldId, &bodyDef);
-
-				s2CreatePolygonShape(bodyId, &s2_defaultShapeDef, &box);
-
-				x += 2.0f * fraction * extent;
-			}
-
-			y += 2.0f * fraction * extent;
-		}
-	}
-
-	static Sample* Create(const Settings& settings, s2SolverType solverType)
-	{
-		return new FarRecovery(settings, solverType);
-	}
-};
-
-static int sampleFarRecovery = RegisterSample("Contact", "Far Recovery", FarRecovery::Create);
