+++ samples/collection/sample_joints.cpp
@@ -531,113 +531,3 @@ public:
 };
 
 static int sampleJointGrid = RegisterSample("Joints", "Joint Grid", JointGrid::Create);
-
-class FarRagdoll : public Sample
-{
-public:
-	FarRagdoll(const Settings& settings, s2SolverType solverType)
-		: Sample(settings, solverType)
-	{
-		//s2Vec2 origin = {0.0f, 0.0f};
-		//s2Vec2 origin = {20000.0f, -15000.0f};
-		s2Vec2 origin = {30000.0f, -25000.0f};
-		// s2Vec2 origin = {900000.0f, -800000.0f};
-
-		if (settings.restart == false)
-		{
-			g_camera.m_center = s2Add({0.0f, 1.0f}, origin);
-			g_camera.m_zoom = 0.05f;
-		}
-
-		{
-			s2BodyDef bodyDef = s2_defaultBodyDef;
-			bodyDef.position = s2Add({0.0f, -1.0f}, origin);
-			s2BodyId groundId = s2CreateBody(m_worldId, &bodyDef);
-			s2Polygon box = s2MakeBox(20.0f, 1.0f);
-			s2CreatePolygonShape(groundId, &s2_defaultShapeDef, &box);
-		}
-
-		m_human.Spawn(m_worldId, s2Add({0.0f, 1.0f}, origin), 1.0f, 1, nullptr);
-	}
-
-	static Sample* Create(const Settings& settings, s2SolverType solverType)
-	{
-		return new FarRagdoll(settings, solverType);
-	}
-
-	Human m_human;
-};
-
-static int sampleFarRagdoll = RegisterSample("Joints", "Far Ragdoll", FarRagdoll::Create);
-
-class FarChain : public Sample
-{
-public:
-	enum
-	{
-		e_count = 40
-	};
-
-	FarChain(const Settings& settings, s2SolverType solverType)
-		: Sample(settings, solverType)
-	{
-		// s2Vec2 origin = {0.0f, 0.0f};
-		s2Vec2 origin = {40000.0f, -35000.0f};
-		// s2Vec2 origin = {900000.0f, -800000.0f};
-
-		if (settings.restart == false)
-		{
-			g_camera.m_center = s2Add({0.0f, -0.5f}, origin);
-			g_camera.m_zoom = 0.2f;
-		}
-
-		s2BodyId groundId = s2_nullBodyId;
-		{
-			s2BodyDef bodyDef = s2_defaultBodyDef;
-			bodyDef.position = origin;
-			groundId = s2CreateBody(m_worldId, &bodyDef);
-		}
-
-		{
-			float hx = 0.1f;
-			s2Capsule capsule = {{-hx, 0.0f}, {hx, 0.0f}, 0.025f};
-
-			s2ShapeDef shapeDef = s2_defaultShapeDef;
-			shapeDef.density = 20.0f;
-
-			s2RevoluteJointDef jointDef = s2DefaultRevoluteJointDef();
-			jointDef.drawSize = 0.02f;
-
-			s2Vec2 prevLocalPivot = {0.0f, e_count * hx};
-			s2BodyId prevBodyId = groundId;
-			for (int i = 0; i < e_count; ++i)
-			{
-				s2BodyDef bodyDef = s2_defaultBodyDef;
-				bodyDef.type = s2_dynamicBody;
-				bodyDef.position = s2Add({(1.0f + 2.0f * i) * hx, e_count * hx}, origin);
-				bodyDef.linearDamping = 0.1f;
-				bodyDef.angularDamping = 0.1f;
-
-				s2BodyId bodyId = s2CreateBody(m_worldId, &bodyDef);
-				s2CreateCapsuleShape(bodyId, &shapeDef, &capsule);
-
-				s2Vec2 pivot = {(2.0f * i) * hx, e_count * hx};
-				jointDef.bodyIdA = prevBodyId;
-				jointDef.bodyIdB = bodyId;
-				jointDef.localAnchorA = prevLocalPivot;
-				jointDef.localAnchorB = {-hx, 0.0f};
-				s2CreateRevoluteJoint(m_worldId, &jointDef);
-
-				prevLocalPivot = {hx, 0.0f};
-				prevBodyId = bodyId;
-			}
-		}
-	}
-
-	static Sample* Create(const Settings& settings, s2SolverType solverType)
-	{
-		return new FarChain(settings, solverType);
-	}
-};
-
-static int sampleFarChain = RegisterSample("Joints", "Far Chain", FarChain::Create);
